Staging

See merge request aloes/device-manager!10

+++ src/boot/11-init-cache.js
@@ -1,12 +1,25 @@
+/* Copyright 2019 Edouard Maleix, read LICENSE */
+
+import logger from '../services/logger';
+
 module.exports = async function initCache(app) {
   try {
+    if (process.env.CLUSTER_MODE) {
+      if (process.env.PROCESS_ID !== '0') return null;
+      if (process.env.INSTANCES_PREFIX && process.env.INSTANCES_PREFIX !== '1') return null;
+    }
+    logger.publish(2, 'loopback', 'boot:initCache:req', '');
+    await app.models.Client.deleteAll();
+    await app.models.SensorResource.deleteAll();
+    // console.log('connected devices', await app.models.Device.find({ 'clients.length': { gt: 0 } }));
+    await app.models.Device.updateAll(
+      // { 'clients.length': { gt: 0 } },
+      { status: false, clients: [] },
+    );
     const devices = await app.models.Device.syncCache('DOWN');
-    const clients = await app.models.Client.deleteAll();
-    // todo keep schedulers cache and find timers
-    const schedulers = await app.models.Scheduler.deleteAll();
-    //  await app.models.Device.updateAll({ 'clients.length': {gt : 0}  }, { clients: [] });
-    return { devices, clients, schedulers };
+    return { devices };
   } catch (error) {
-    throw error;
+    logger.publish(2, 'loopback', 'boot:initCache:err', error);
+    return null;
   }
 };
